apps/discord-bot/pyproject.toml
[project]
name = "assetbridge-discord-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "sqlalchemy>=2.0",
    "matplotlib>=3.9",
    "numpy>=1.26",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "litellm>=1.40",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
